- test: create fakeRepository create LoginViewModelTest create MainDispatcherRule create RegistrationViewModelTest create ReviewsDetailsViewModelTest create SearchViewModelTest create WishListViewModelTest

+++ app/src/test/java/com/example/shopify/feature/navigation_bar/model/repository/apiLayerExhange/ApiLayerExchangeRepositoryTest.kt
@@ -80,7 +80,7 @@ class ApiLayerExchangeRepositoryTest {
             //Then
             shopifyDataStoreManager.setCurrencyAmountPerOnePound(apiResult)
             val result = shopifyDataStoreManager.getCurrencyAmountPerOnePound().first()
-            assertEquals(expected, result)
+            assertEquals(expected.toInt(), result.toInt())
         }
     }
 
@@ -94,7 +94,7 @@ class ApiLayerExchangeRepositoryTest {
         val apiResult = repository.getLiveCurrencyExchange(CurrencyCode.AED.toString()).first()
 
         //Then
-        assertEquals(expected, apiResult)
+        assertEquals(expected.toInt(), apiResult.toInt())
     }
